# usm: Improve connection creation to reduce allocations