@guyarb 2023-04-17
> Do we have a performance/pprof report for this change to see the actual improvement?

added to the PR description